move DashboardMetadataDAOImpl class to internal.database package
