+++ components/dashboards/org.wso2.carbon.dashboards.core/findbugs-exclude.xml
@@ -17,7 +17,7 @@
 
 <FindBugsFilter>
     <Match>
-        <Class name="org.wso2.carbon.dashboards.core.internal.dao.impl.DashboardMetadataDAOImpl"/>
+        <Class name="org.wso2.carbon.dashboards.core.internal.database.DashboardMetadataDAOImpl"/>
         <Bug pattern="SQL_PREPARED_STATEMENT_GENERATED_FROM_NONCONSTANT_STRING"/>
     </Match>
 </FindBugsFilter>
